Removing aria-atomic on regions
Removing the aria-atomic attribute due to a really annoying bug in IE10
running in compatibility mode, where it apparently doesn't like certain
unspec'd attributes added to elements.  Need to read up on expando
elements, this bug that Microsoft has graciously decided to not fix
until IE11, and if a newer version of jQuery will like this better.